--- Day3/code/src/Router2.js
@@ -1,11 +1,0 @@
-import React from "react";
-import Effect1 from "./Effect1";
-
-export default function Router2() {
-  return (
-    <>
-      <h1>Effect1</h1>
-      <Effect1 />
-    </>
-  );
-}
