--- Day3/code/src/Router3.js
@@ -1,11 +1,0 @@
-import React from "react";
-import Navi1 from "./Navi1";
-
-export default function Router3() {
-  return (
-    <>
-      <h1>Navi1</h1>
-      <Navi1 />
-    </>
-  );
-}
