--- Day3/code/src/State1.js
@@ -1,18 +1,0 @@
-import React, { useState } from 'react';
-
-function State() {
-  const [name, setName] = useState(false);
-
-  const onClick = () => {
-    setName(true);
-  };
-
-  return (
-    <>
-      <p>{name ? 'true' : 'false'}</p>
-      <button onClick={onClick}>go!</button>
-    </>
-  );
-}
-
-export default State;
