--- Day3/code/src/State3.js
@@ -1,20 +1,0 @@
-import React, { useState } from "react";
-
-function Counter() {
-  const [count, setCount] = useState(0);
-
-  function handleAlertClick() {
-    setTimeout(() => {
-      alert("You clicked on: " + count);
-    }, 3000);
-  }
-  return (
-    <div>
-      <p>You clicked {count} times</p>
-      <button onClick={() => setCount(count + 1)}>Click me</button>
-      <button onClick={handleAlertClick}>Show alert</button>
-    </div>
-  );
-}
-
-export default Counter;
